+++ clients/admin-ui/src/pages/config-wizard/index.tsx
@@ -1,11 +1,15 @@
 import type { NextPage } from "next";
 import { useState } from "react";
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import { Flags } from "react-feature-flags";
 
 import Header from "~/features/common/Header";
 import Layout from "~/features/common/Layout";
 import ConfigWizardWalkthrough from "~/features/config-wizard/ConfigWizardWalkthrough";
 
 import Setup from "../../features/config-wizard/setup";
+import Custom404 from "../404";
 
 const ConfigWizard: NextPage = () => {
   const [configWizardStep, setConfigWizardStep] = useState(false);
@@ -15,14 +19,26 @@ const ConfigWizard: NextPage = () => {
   };
 
   return !configWizardStep ? (
-    <Layout title="Config Wizard" noPadding>
-      <Setup wizardStep={handleWizardStep} />
-    </Layout>
+    <Flags
+      authorizedFlags={["configWizardFlag"]}
+      renderOn={() => (
+        <Layout title="Config Wizard" noPadding>
+          <Setup wizardStep={handleWizardStep} />
+        </Layout>
+      )}
+      renderOff={() => <Custom404 />}
+    />
   ) : (
-    <>
-      <Header />
-      <ConfigWizardWalkthrough />
-    </>
+    <Flags
+      authorizedFlags={["configWizardFlag"]}
+      renderOn={() => (
+        <>
+          <Header />
+          <ConfigWizardWalkthrough />
+        </>
+      )}
+      renderOff={() => <Custom404 />}
+    />
   );
 };